+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <!-- 
-  Catena-X - Digital Product Passport Application 
+  Tractus-X - Digital Product Passport Application 
  
   Copyright (c) 2022, 2024 BASF SE, BMW AG, Henkel AG & Co. KGaA
   Copyright (c) 2022, 2024 Contributors to the Eclipse Foundation
@@ -25,6 +25,11 @@
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [released]
+## [v2.1.3] - 19-02-2024
+
+## Updated
+- Renamed "Catena-X" to "Tractus-X" in header license prefix
 
 ## [released]
 ## [v2.1.2] - 16-02-2024